+++ client/components/recurring-transactions/new-recurring-transaction.tsx
@@ -1,14 +1,16 @@
-import React, { useCallback, useState } from 'react';
+import React, { useCallback, useState, useContext } from 'react';
 import { useHistory, useParams } from 'react-router-dom';
+import moment from 'moment';
 
 import {
+    assert,
     notify,
     translate as $t,
     UNKNOWN_TRANSACTION_TYPE,
     noValueFoundMessage,
 } from '../../helpers';
 
-import URL from './urls';
+import URL from '../../urls';
 
 import { Form, ValidatedTextInput, BackLink } from '../ui';
 import AmountInput from '../ui/amount-input';
@@ -17,18 +19,22 @@ import MultipleSelect from '../ui/multiple-select';
 
 import TypeSelect from '../reports/type-select';
 
-import moment from 'moment';
+import { ViewContext } from '../drivers';
+
 import { createRecurringTransaction } from '../../store/backend';
 
 export default () => {
+    const currentView = useContext(ViewContext);
+    assert(!!currentView.account, 'Account not provided to view');
+
     const {
-        accountId: accountIdStr,
+        value: accountIdStr,
         label: rawPredefinedLabel,
         amount: rawPredefinedAmount,
         day: rawPredefinedDay,
         type: predefinedType,
     } = useParams<{
-        accountId: string;
+        value: string;
         label?: string;
         amount?: string;
         day?: string;
@@ -56,9 +62,8 @@ export default () => {
         }
     }
 
+    const listUrl = URL.recurringTransactions.url(currentView.driver);
     const accountId = Number.parseInt(accountIdStr, 10);
-
-    const listUrl = URL.listAccountRecurringTransactions(accountId);
 
     const daysList = [];
     for (let i = 1; i <= 31; ++i) {