+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Emails are now obfuscated in logs.
 - Onboarding shows more obviously it's possible to use manual accounts with Kresus.
+- Recurring transactions are now available from the lateral menu.
 
 ### Bug fixes